+++ package/src/cookie_consent/client/cookie_consent_provider.tsx
@@ -54,7 +54,7 @@ export default function CookieConsentProvider({ requiresConsent = true, children
     requiresConsent?: boolean;
     children: React.ReactNode;
 }): React.ReactElement {
-    const { consentCookieName, dateCookieName, maxAge, policyDate, privacyPolicyPath } = useMemo(() => {
+    const { consentCookieName, dateCookieName, maxAge, policyDate, privacyPolicyPath, showPrivacyPolicy } = useMemo(() => {
         const cc = requireCookieConsentConfig(config.cookieConsent);
         return {
             consentCookieName: cc.consentCookieName ?? cookieConsentCookieKey,
@@ -62,6 +62,7 @@ export default function CookieConsentProvider({ requiresConsent = true, children
             maxAge: cc.cookieMaxAge ?? 31536000,
             policyDate: cc.privacyPolicyDate ? new Date(cc.privacyPolicyDate) : null,
             privacyPolicyPath: cc.privacyPolicyPath ?? '/privacy-policy',
+            showPrivacyPolicy: cc.showPrivacyPolicy ?? true,
         };
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
@@ -131,8 +132,8 @@ export default function CookieConsentProvider({ requiresConsent = true, children
     }, [pathname, privacyPolicyUpdated, privacyPolicyPath, acknowledgePrivacyPolicyUpdate]);
 
     const contextValue = useMemo(
-        () => ({ consent, requiresConsent, privacyPolicyUpdated, isMounted, setConsent, acknowledgePrivacyPolicyUpdate, privacyPolicyPath }),
-        [consent, requiresConsent, privacyPolicyUpdated, isMounted, setConsent, acknowledgePrivacyPolicyUpdate, privacyPolicyPath],
+        () => ({ consent, requiresConsent, privacyPolicyUpdated, isMounted, setConsent, acknowledgePrivacyPolicyUpdate, privacyPolicyPath, showPrivacyPolicy }),
+        [consent, requiresConsent, privacyPolicyUpdated, isMounted, setConsent, acknowledgePrivacyPolicyUpdate, privacyPolicyPath, showPrivacyPolicy],
     );
 
     return (