+++ package/src/server/components/server_provider_static.tsx
@@ -105,6 +105,7 @@ export default async function LocationzationProvider({ language, messages, child
                 config.cookieConsent.gdprCountries,
                 config.errorHandling,
                 config.cookieConsent.countryHeaderNames,
+                config.generate,
             )
             : false;